# api/docs/swagger.yaml
basePath: /
definitions:
  api.apiError:
    properties:
      code:
        type: string
      message:
        type: string
      param:
        type: string
      type:
        type: string
    type: object
  api.apiErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.apiError'
      object:
        description: '"error"'
        type: string
    type: object
  api.changePasswordRequest:
    properties:
      new_password:
        example: newsecret
        type: string
      reset_token:
        example: a3f1...
        type: string
    type: object
...
  api.deletedResponse:
    properties:
      deleted:
        type: boolean
      id:
        type: string
      object:
        type: string
    type: object
...
  api.documentRequest:
    properties:
      author:
        example: Парламент
        type: string
      country:
        example: by
        type: string
      date:
        example: "1999-07-26"
        type: string
      name:
        example: Трудовой кодекс
        type: string
      number:
        example: 296-З
        type: string
      subtype:
        example: закон
        type: string
    type: object
  api.documentResponse:
    properties:
      author:
        type: string
      complete:
        type: boolean
      country:
        type: string
      created:
        type: integer
      date:
        type: string
      id:
        type: string
      name:
        type: string
      npa_level:
        type: integer
      number:
        type: string
      object:
        type: string
      subtype:
        type: string
      user_id:
        type: string
    type: object
...
  api.listResponse-api_documentResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/api.documentResponse'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: string
      object:
        description: '"list"'
        type: string
    type: object
...
  api.passwordResetRequest:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
  api.passwordResetResponse:
    properties:
      object:
        description: '"token.password_reset"'
        type: string
      reset_token:
        type: string
    type: object
...
host: legist.nadevko.cc
info:
  contact: {}
  description: |-
    AI assistant for comparing revisions of normative legal acts (НПА) of the Republic of Belarus.

    ## Authentication
    All endpoints except public auth routes require a Bearer token in the `Authorization` header.
    Obtain tokens via `POST /sessions`. Refresh via `POST /tokens/refresh`.

    ## Request IDs
    Every response includes a `Request-Id` header for debugging.

    ## Idempotency
    `POST` requests require an `Idempotency-Key` header.
    Repeating a request with the same key returns the cached response.
    Keys expire after 24 hours. Reusing a key for a different endpoint returns 422.

    ## Roles
    Users have `role`: `user` or `admin` (in `users` table and JWT `role` claim).
    Only `admin` may change another user's role via `PATCH /users/:id`.
    In dev (`ENV=dev`), new registrations default to `admin`; otherwise default is `user`.

    ## Ownership and access control
    Resources belong to the authenticated user. Attempting to read or mutate
    another user's private resource returns 404 (not 403) to avoid leaking existence.
    Public resources (`user_id IS NULL`) are readable by any authenticated user but cannot be mutated.
    List `owner` query (Stripe-like): omit → non-admin sees only own rows; admin sees own + public.
    `owner=null` or `owner=public` (aliases, admin only) → public rows only. `owner=<your user id>` → own rows only. Other ids are rejected.

...

...

...

...

...

...

...

...

...

...
paths:
...
  /documents:
    get:
      parameters:
      - description: 'omit: own (user) or own+public (admin); null|public=public only
          (admin); or your user id'
        in: query
        name: owner
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: starting_after
        type: string
      - description: Cursor
        in: query
        name: ending_before
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.listResponse-api_documentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: List documents
      tags:
      - Documents
    post:
      consumes:
      - application/json
      parameters:
      - description: Document fields (all optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.documentRequest'
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.documentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "409":
          description: Duplicate subtype+number+date
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a document
      tags:
      - Documents
  /documents/{id}:
    delete:
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.deletedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete document
      tags:
      - Documents
    get:
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - collectionFormat: csv
        description: 'Expand: files'
        in: query
        items:
          type: string
        name: expand[]
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.documentResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Get document
      tags:
      - Documents
    patch:
      consumes:
      - application/json
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.documentRequest'
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.documentResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Update document Work-level fields
      tags:
      - Documents
...
  /tokens/password-reset:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Reset token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.deletedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      summary: Change password using reset token
      tags:
      - Sessions
    post:
      consumes:
      - application/json
      description: Always returns 200 regardless of whether the email exists,
      parameters:
      - description: Email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.passwordResetRequest'
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.passwordResetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      summary: Request password reset token
      tags:
      - Sessions
...
securityDefinitions:
  BearerAuth:
    description: 'Enter token in format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...